# ARROW-10216: [Rust] Simd implementation for primitive min/max kernels