--- a/rust/arrow/src/compute/kernels/aggregate.rs
+++ b/rust/arrow/src/compute/kernels/aggregate.rs
@@ -57,6 +57,7 @@ fn min_max_string<T: StringOffsetSizeTrait, F: Fn(&str, &str) -> bool>(
 }
 
 /// Returns the minimum value in the array, according to the natural order.
+#[cfg(not(all(any(target_arch = "x86", target_arch = "x86_64"), feature = "simd")))]
 pub fn min<T>(array: &PrimitiveArray<T>) -> Option<T::Native>
 where
     T: ArrowNumericType,
@@ -179,6 +180,181 @@ where
     }
 }
 
+#[cfg(all(any(target_arch = "x86", target_arch = "x86_64"), feature = "simd"))]
+mod simd {
+    use crate::array::{Array, PrimitiveArray};
+    use crate::datatypes::ArrowNumericType;
+    use std::marker::PhantomData;
+    use std::ops::Add;
+    use std::ops::BitAnd;
+
+    pub(super) trait SimdAggregate<T: ArrowNumericType> {
+        fn init_accumulator_scalar() -> T::Native;
+
+        fn init_accumulator_chunk() -> T::Simd {
+            T::init(Self::init_accumulator_scalar())
+        }
+
+        fn accumulate_chunk_non_null(accumulator: &mut T::Simd, chunk: T::Simd);
+        fn accumulate_chunk_nullable(
+            accumulator: &mut T::Simd,
+            chunk: T::Simd,
+            mask: T::SimdMask,
+        );
+
+        fn accumulate_scalar(accumulator: &mut T::Native, value: T::Native);
+
+        fn reduce(accumulator: T::Simd) -> T::Native {
+            // reduce by first writing to a temporary and then use scalar operations
+            // this should be about the same performance as extracting individual lanes
+            // but allows us to reuse the scalar reduction logic
+            let tmp = &mut [T::default_value(); 64];
+            T::write(accumulator, &mut tmp[0..T::lanes()]);
+
+            let mut reduced = Self::init_accumulator_scalar();
+            tmp[0..T::lanes()]
+                .iter()
+                .for_each(|value| Self::accumulate_scalar(&mut reduced, *value));
+
+            reduced
+        }
+    }
+
+    pub(super) struct SumAggregate<T: ArrowNumericType> {
+        default_value: T::Native,
+    }
+
+    impl<T: ArrowNumericType> SimdAggregate<T> for SumAggregate<T>
+    where
+        T::Native: Add<Output = T::Native>,
+    {
+        fn init_accumulator_scalar() -> T::Native {
+            T::default_value()
+        }
+
+        fn accumulate_chunk_non_null(accumulator: &mut T::Simd, chunk: T::Simd) {
+            *accumulator = *accumulator + chunk;
+        }
+
+        fn accumulate_chunk_nullable(
+            accumulator: &mut T::Simd,
+            chunk: T::Simd,
+            vecmask: T::SimdMask,
+        ) {
+            let zero = T::init(T::default_value());
+            let blended = T::mask_select(vecmask, chunk, zero);
+
+            *accumulator = *accumulator + blended;
+        }
+
+        fn accumulate_scalar(accumulator: &mut T::Native, value: T::Native) {
+            *accumulator = *accumulator + value
+        }
+    }
+
+    pub(super) struct MinAggregate<T: ArrowNumericType> {
+        default_value: T::Native,
+    }
+
+    impl<T: ArrowNumericType> SimdAggregate<T> for MinAggregate<T>
+    where
+        T::Native: PartialOrd,
+    {
+        fn init_accumulator_scalar() -> T::Native {
+            T::identity_for_min_op()
+        }
+
+        fn accumulate_chunk_non_null(accumulator: &mut T::Simd, chunk: T::Simd) {
+            let cmp_mask = T::lt(chunk, *accumulator);
+
+            *accumulator = T::mask_select(cmp_mask, chunk, *accumulator);
+        }
+
+        fn accumulate_chunk_nullable(
+            accumulator: &mut T::Simd,
+            chunk: T::Simd,
+            vecmask: T::SimdMask,
+        ) {
+            let cmp_mask = T::lt(chunk, *accumulator);
+            let blend_mask = T::mask_and(vecmask, cmp_mask);
+
+            *accumulator = T::mask_select(blend_mask, chunk, *accumulator);
+        }
+
+        fn accumulate_scalar(accumulator: &mut T::Native, value: T::Native) {
+            if value < *accumulator {
+                *accumulator = value
+            }
+        }
+    }
+
+    pub(super) fn simd_aggregation<T: ArrowNumericType, A: SimdAggregate<T>>(
+        array: &PrimitiveArray<T>,
+    ) -> Option<T::Native> {
+        let null_count = array.null_count();
+
+        if null_count == array.len() {
+            return None;
+        }
+
+        let data: &[T::Native] = array.value_slice(0, array.len());
+
+        let mut chunk_acc = A::init_accumulator_chunk();
+        let mut rem_acc = A::init_accumulator_scalar();
+
+        match array.data().null_buffer() {
+            None => {
+                let data_chunks = data.chunks_exact(64);
+                let remainder = data_chunks.remainder();
+
+                data_chunks.for_each(|chunk| {
+                    chunk.chunks_exact(T::lanes()).for_each(|chunk| {
+                        let chunk = T::load(&chunk);
+                        A::accumulate_chunk_non_null(&mut chunk_acc, chunk);
+                    });
+                });
+
+                remainder.iter().for_each(|value| {
+                    A::accumulate_scalar(&mut rem_acc, *value);
+                });
+            }
+            Some(buffer) => {
+                // process data in chunks of 64 elements since we also get 64 bits of validity information at a time
+                let data_chunks = data.chunks_exact(64);
+                let remainder = data_chunks.remainder();
+
+                let bit_chunks = buffer.bit_chunks(array.offset(), array.len());
+                let remainder_bits = bit_chunks.remainder_bits();
+
+                data_chunks.zip(bit_chunks).for_each(|(chunk, mut mask)| {
+                    // split chunks further into slices corresponding to the vector length
+                    // the compiler is able to unroll this inner loop and remove bounds checks
+                    // since the outer chunk size (64) is always a multiple of the number of lanes
+                    chunk.chunks_exact(T::lanes()).for_each(|chunk| {
+                        let vecmask = T::mask_from_u64(mask);
+                        let chunk = T::load(&chunk);
+
+                        A::accumulate_chunk_nullable(&mut chunk_acc, chunk, vecmask);
+
+                        mask = mask >> T::lanes();
+                    });
+                });
+
+                remainder.iter().enumerate().for_each(|(i, value)| {
+                    if remainder_bits & (1 << i) != 0 {
+                        A::accumulate_scalar(&mut rem_acc, *value)
+                    }
+                });
+            }
+        }
+
+        let mut total = A::reduce(chunk_acc);
+        A::accumulate_scalar(&mut total, rem_acc);
+
+        Some(total)
+    }
+}
+
 /// Returns the sum of values in the array.
 ///
 /// Returns `None` if the array is empty or only contains null values.
@@ -187,79 +363,19 @@ pub fn sum<T: ArrowNumericType>(array: &PrimitiveArray<T>) -> Option<T::Native>
 where
     T::Native: Add<Output = T::Native>,
 {
-    let null_count = array.null_count();
+    use simd::*;
 
-    if null_count == array.len() {
-        return None;
-    }
+    simd::simd_aggregation::<T, SumAggregate<T>>(&array)
+}
 
-    let data: &[T::Native] = array.value_slice(0, array.len());
+#[cfg(all(any(target_arch = "x86", target_arch = "x86_64"), feature = "simd"))]
+pub fn min<T: ArrowNumericType>(array: &PrimitiveArray<T>) -> Option<T::Native>
+where
+    T::Native: PartialOrd,
+{
+    use simd::*;
 
-    let mut vector_sum = T::init(T::default_value());
-    let mut remainder_sum = T::default_value();
-
-    match array.data().null_buffer() {
-        None => {
-            let data_chunks = data.chunks_exact(64);
-            let remainder = data_chunks.remainder();
-
-            data_chunks.for_each(|chunk| {
-                chunk.chunks_exact(T::lanes()).for_each(|chunk| {
-                    let chunk = T::load(&chunk);
-                    vector_sum = vector_sum + chunk;
-                });
-            });
-
-            remainder.iter().for_each(|value| {
-                remainder_sum = remainder_sum + *value;
-            });
-        }
-        Some(buffer) => {
-            // process data in chunks of 64 elements since we also get 64 bits of validity information at a time
-            let data_chunks = data.chunks_exact(64);
-            let remainder = data_chunks.remainder();
-
-            let bit_chunks = buffer.bit_chunks(array.offset(), array.len());
-            let remainder_bits = bit_chunks.remainder_bits();
-
-            data_chunks.zip(bit_chunks).for_each(|(chunk, mut mask)| {
-                // split chunks further into slices corresponding to the vector length
-                // the compiler is able to unroll this inner loop and remove bounds checks
-                // since the outer chunk size (64) is always a multiple of the number of lanes
-                chunk.chunks_exact(T::lanes()).for_each(|chunk| {
-                    let zero = T::init(T::default_value());
-                    let vecmask = T::mask_from_u64(mask);
-                    let chunk = T::load(&chunk);
-                    let blended = T::mask_select(vecmask, chunk, zero);
-
-                    vector_sum = vector_sum + blended;
-
-                    mask = mask >> T::lanes();
-                });
-            });
-
-            remainder.iter().enumerate().for_each(|(i, value)| {
-                if remainder_bits & (1 << i) != 0 {
-                    remainder_sum = remainder_sum + *value;
-                }
-            });
-        }
-    }
-
-    // calculate horizontal sum of accumulator by writing to a temporary
-    // this is probably faster than extracting individual lanes
-    // the compiler is free to optimize this to something faster
-    let tmp = &mut [T::default_value(); 64];
-    T::write(vector_sum, &mut tmp[0..T::lanes()]);
-
-    let mut total_sum = T::default_value();
-    tmp[0..T::lanes()]
-        .iter()
-        .for_each(|lane| total_sum = total_sum + *lane);
-
-    total_sum = total_sum + remainder_sum;
-
-    Some(total_sum)
+    simd::simd_aggregation::<T, MinAggregate<T>>(&array)
 }
 
 #[cfg(test)]