@jhorstmann 2020-11-18
I still need to add a few more tests involving infinities and NaNs