@jhorstmann 2020-11-20
Both scalar and simd implementations have issues involving null values, I added `ignore` tests with comments. The handling of NaN is also not consistent with the sort kernel, which follows the postgres model of sorting NaN as larger than any other number (including infinity). Solving those issues could be quite tricky and degrade the performance again.