@Dandandan 2020-12-08
@jhorstmann 
I wonder if we should support a standardized ordering like  IEEE 754 instead?

There is a (unstable) implementation in rust std for it.
https://doc.rust-lang.org/beta/std/primitive.f64.html#method.total_cmp

Sounds something we could have in the source for now (it is only a few lines), I think it might also simplify some code like the current sort kernels? What do you think?