@jhorstmann 2020-12-09
@Dandandan that sounds like good idea as it is a bit more standardized. The current ordering was choosen initially for the sort kernels, as it is the same ordering that Postgres implements. The main difference seems to be the handling of negative NaN. If we decide on a different ordering we should keep sort and min/max consistent.

One nice thing about the current impl for min/max is that all numeric types are handled the same, with the compiler optimizing away the `is_nan` function for integer types.